renames add_member to member_fields
